+++ jaeger_client/config.py
@@ -1,27 +1,23 @@
 # Copyright (c) 2016 Uber Technologies, Inc.
 #
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
 #
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
+# http://www.apache.org/licenses/LICENSE-2.0
 #
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-# THE SOFTWARE.
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 from __future__ import absolute_import
 from builtins import object
 
+from builtins import object
 import logging
+import os
 import threading
 
 import opentracing
@@ -48,9 +44,10 @@ from .constants import (
     BAGGAGE_HEADER_PREFIX,
     DEBUG_ID_HEADER_KEY,
 )
-from .metrics import Metrics
+from .metrics import LegacyMetricsFactory, MetricsFactory, Metrics
 from .utils import get_boolean, ErrorReporter
 
+DEFAULT_REPORTING_HOST = 'localhost'
 DEFAULT_REPORTING_PORT = 5775
 DEFAULT_SAMPLING_PORT = 5778
 LOCAL_AGENT_DEFAULT_ENABLED = True
@@ -82,24 +79,26 @@ class Config(object):
     _initialized = False
     _initialized_lock = threading.Lock()
 
-    def __init__(self, config, metrics=None, service_name=None):
+    def __init__(self, config, metrics=None, service_name=None, metrics_factory=None):
         """
-        :param metrics: an instance of Metrics class, or None
+        :param metrics: an instance of Metrics class, or None. This parameter
+            has been deprecated, please use metrics_factory instead.
         :param service_name: default service name.
             Can be overwritten by config['service_name'].
+        :param metrics_factory: an instance of MetricsFactory class, or None.
         """
         self.config = config
         if get_boolean(self.config.get('metrics', True), True):
-            self._metrics = metrics or Metrics()
+            self._metrics_factory = metrics_factory or LegacyMetricsFactory(metrics or Metrics())
         else:
             # if metrics are explicitly disabled, use a dummy
-            self._metrics = Metrics()
+            self._metrics_factory = MetricsFactory()
         self._service_name = config.get('service_name', service_name)
         if self._service_name is None:
             raise ValueError('service_name required in the config or param')
 
         self._error_reporter = ErrorReporter(
-            metrics=self.metrics,
+            metrics=Metrics(),
             logger=logger if self.logging else None,
         )
 
@@ -214,8 +213,30 @@ class Config(object):
             return DEFAULT_REPORTING_PORT
 
     @property
+    def local_agent_reporting_host(self):
+        # noinspection PyBroadException
+        try:
+            return self.local_agent_group()['reporting_host']
+        except:
+            return DEFAULT_REPORTING_HOST
+
+    @property
     def max_operations(self):
         return self.config.get('max_operations', None)
+
+    @property
+    def tags(self):
+        """
+        :return: Returns tags from config and `JAEGER_TAGS` environment variable
+        to use as process-wide tracer tags
+        """
+        tags = self.config.get('tags', {})
+        env_tags = os.environ.get('JAEGER_TAGS', '')
+        if env_tags:
+            for kv in env_tags.split(','):
+                key, value = kv.split('=')
+                tags[key.strip()] = value.strip()
+        return tags
 
     @staticmethod
     def initialized():
@@ -242,7 +263,7 @@ class Config(object):
                 channel=channel,
                 service_name=self.service_name,
                 logger=logger,
-                metrics=self.metrics,
+                metrics_factory=self._metrics_factory,
                 error_reporter=self.error_reporter,
                 sampling_refresh_interval=self.sampling_refresh_interval,
                 max_operations=self.max_operations)
@@ -254,7 +275,7 @@ class Config(object):
             batch_size=self.reporter_batch_size,
             flush_interval=self.reporter_flush_interval,
             logger=logger,
-            metrics=self.metrics,
+            metrics_factory=self._metrics_factory,
             error_reporter=self.error_reporter)
 
         if self.logging:
@@ -273,10 +294,11 @@ class Config(object):
             service_name=self.service_name,
             reporter=reporter,
             sampler=sampler,
-            metrics=self.metrics,
+            metrics_factory=self._metrics_factory,
             trace_id_header=self.trace_id_header,
             baggage_header_prefix=self.baggage_header_prefix,
             debug_id_header=self.debug_id_header,
+            tags=self.tags,
         )
 
     def _initialize_global_tracer(self, tracer):
@@ -294,7 +316,7 @@ class Config(object):
         """
         logger.info('Initializing Jaeger Tracer with UDP reporter')
         return LocalAgentSender(
-            host='localhost',
+            host=self.local_agent_reporting_host,
             sampling_port=self.local_agent_sampling_port,
             reporting_port=self.local_agent_reporting_port,
             io_loop=io_loop