+++ jaeger_client/constants.py
@@ -1,27 +1,22 @@
 # Copyright (c) 2016 Uber Technologies, Inc.
 #
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
 #
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
+# http://www.apache.org/licenses/LICENSE-2.0
 #
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-# THE SOFTWARE.
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 from __future__ import absolute_import, unicode_literals, print_function
 
 from . import __version__
 
+import six
 
 # Max number of bits to use when generating random ID
 MAX_ID_BITS = 64
@@ -33,10 +28,10 @@ DEFAULT_SAMPLING_INTERVAL = 60
 DEFAULT_FLUSH_INTERVAL = 1
 
 # Name of the HTTP header used to encode trace ID
-TRACE_ID_HEADER = 'uber-trace-id'
+TRACE_ID_HEADER = 'uber-trace-id' if six.PY3 else b'uber-trace-id'
 
 # Prefix for HTTP headers used to record baggage items
-BAGGAGE_HEADER_PREFIX = 'uberctx-'
+BAGGAGE_HEADER_PREFIX = 'uberctx-' if six.PY3 else b'uberctx-'
 
 # The name of HTTP header or a TextMap carrier key which, if found in the
 # carrier, forces the trace to be sampled as "debug" trace. The value of the