+++ jaeger_client/reporter.py
@@ -1,22 +1,16 @@
 # Copyright (c) 2016 Uber Technologies, Inc.
 #
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
 #
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
+# http://www.apache.org/licenses/LICENSE-2.0
 #
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-# THE SOFTWARE.
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 from __future__ import absolute_import
 from builtins import object
@@ -31,7 +25,7 @@ from concurrent.futures import Future
 from .constants import DEFAULT_FLUSH_INTERVAL
 from . import thrift
 from . import ioloop_util
-from .metrics import Metrics
+from .metrics import Metrics, LegacyMetricsFactory
 from .utils import ErrorReporter
 
 from thrift.protocol import TCompactProtocol
@@ -80,7 +74,8 @@ class Reporter(NullReporter):
     """Receives completed spans from Tracer and submits them out of process."""
     def __init__(self, channel, queue_capacity=100, batch_size=10,
                  flush_interval=DEFAULT_FLUSH_INTERVAL, io_loop=None,
-                 error_reporter=None, metrics=None, **kwargs):
+                 error_reporter=None, metrics=None, metrics_factory=None,
+                 **kwargs):
         """
         :param channel: a communication channel to jaeger-agent
         :param queue_capacity: how many spans we can hold in memory before
@@ -90,7 +85,9 @@ class Reporter(NullReporter):
         :param io_loop: which IOLoop to use. If None, try to get it from
             channel (only works if channel is tchannel.sync)
         :param error_reporter:
-        :param metrics:
+        :param metrics: an instance of Metrics class, or None. This parameter
+            has been deprecated, please use metrics_factory instead.
+        :param metrics_factory: an instance of MetricsFactory class, or None.
         :param kwargs:
             'logger'
         :return:
@@ -100,8 +97,9 @@ class Reporter(NullReporter):
         self._channel = channel
         self.queue_capacity = queue_capacity
         self.batch_size = batch_size
-        self.metrics = metrics or Metrics()
-        self.error_reporter = error_reporter or ErrorReporter(self.metrics)
+        self.metrics_factory = metrics_factory or LegacyMetricsFactory(metrics or Metrics())
+        self.metrics = ReporterMetrics(self.metrics_factory)
+        self.error_reporter = error_reporter or ErrorReporter(Metrics())
         self.logger = kwargs.get('logger', default_logger)
         self.agent = Agent.Client(self._channel, self)
 
@@ -133,11 +131,11 @@ class Reporter(NullReporter):
             with self.stop_lock:
                 stopped = self.stopped
             if stopped:
-                self.metrics.count(Metrics.REPORTER_DROPPED, 1)
+                self.metrics.reporter_dropped(1)
             else:
                 self.queue.put_nowait(span)
         except tornado.queues.QueueFull:
-            self.metrics.count(Metrics.REPORTER_DROPPED, 1)
+            self.metrics.reporter_dropped(1)
 
     @tornado.gen.coroutine
     def _consume_queue(self):
@@ -183,14 +181,14 @@ class Reporter(NullReporter):
         try:
             spans = thrift.make_zipkin_spans(spans)
             yield self._send(spans)
-            self.metrics.count(Metrics.REPORTER_SUCCESS, len(spans))
+            self.metrics.reporter_success(len(spans))
         except socket.error as e:
+            self.metrics.reporter_socket(len(spans))
             self.error_reporter.error(
-                Metrics.REPORTER_SOCKET, len(spans),
                 'Failed to submit trace to jaeger-agent socket: %s', e)
         except Exception as e:
+            self.metrics.reporter_failure(len(spans))
             self.error_reporter.error(
-                Metrics.REPORTER_FAILURE, len(spans),
                 'Failed to submit trace to jaeger-agent: %s', e)
 
     @tornado.gen.coroutine
@@ -220,6 +218,18 @@ class Reporter(NullReporter):
     def _flush(self):
         yield self.queue.put(self.stop)
         yield self.queue.join()
+
+
+class ReporterMetrics(object):
+    def __init__(self, metrics_factory):
+        self.reporter_success = \
+            metrics_factory.create_counter(name='jaeger.spans', tags={'reported': 'true'})
+        self.reporter_failure = \
+            metrics_factory.create_counter(name='jaeger.spans', tags={'reported': 'false'})
+        self.reporter_dropped = \
+            metrics_factory.create_counter(name='jaeger.spans', tags={'dropped': 'true'})
+        self.reporter_socket = \
+            metrics_factory.create_counter(name='jaeger.spans', tags={'socket_error': 'true'})
 
 
 class CompositeReporter(NullReporter):