+++ tests/test_span.py
@@ -1,22 +1,16 @@
 # Copyright (c) 2016 Uber Technologies, Inc.
 #
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
 #
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
+# http://www.apache.org/licenses/LICENSE-2.0
 #
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-# THE SOFTWARE.
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 from __future__ import absolute_import
 
@@ -44,6 +38,19 @@ def test_baggage():
     span.set_baggage_item('X_y', '123')
     assert span.get_baggage_item('X_y') == '123'
     assert span.get_baggage_item('x-Y') is None
+
+
+def test_baggage_logs():
+    ctx = SpanContext(trace_id=1, span_id=2, parent_id=None, flags=1)
+    span = Span(context=ctx, operation_name='x', tracer=None)
+    span.set_baggage_item('x', 'a')
+    assert span.get_baggage_item('x') == 'a'
+    assert len(span.logs) == 1
+    assert span.logs[0].value == '{"value": "a", "event": "baggage", "key": "x"}'
+    span.set_baggage_item('x', 'b')  # override
+    assert span.get_baggage_item('x') == 'b'
+    assert len(span.logs) == 2
+    assert span.logs[1].value == '{"override": "true", "value": "b", "event": "baggage", "key": "x"}'
 
 
 def test_sampling_priority(tracer):