+++ tests/test_tracer.py
@@ -1,25 +1,20 @@
 # Copyright (c) 2016 Uber Technologies, Inc.
 #
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
 #
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
+# http://www.apache.org/licenses/LICENSE-2.0
 #
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-# THE SOFTWARE.
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 import mock
 import random
+import six
 
 import six
 import pytest
@@ -198,6 +193,7 @@ def test_tracer_tags_no_hostname():
         'jaeger.hostname': 'dream-host.com',
         'sampler.type': 'const',
         'sampler.param': 'True',
+        'global-tag': 'global-tag'
     }),
     ('child', {
         'jaeger.version': None,
@@ -216,7 +212,10 @@ def test_tracer_tags_on_root_span(span_type, expected_tags):
     reporter = mock.MagicMock()
     sampler = ConstSampler(True)
     with mock.patch('socket.gethostname', return_value='dream-host.com'):
-        tracer = Tracer(service_name='x', reporter=reporter, sampler=sampler)
+        tracer = Tracer(service_name='x',
+                        reporter=reporter,
+                        sampler=sampler,
+                        tags={'global-tag': 'global-tag'})
         span = tracer.start_span(operation_name='root')
         if span_type == 'child':
             span = tracer.start_span('child', child_of=span)